requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
#Django==3.0.3
django-cors-headers==3.4.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
mysqlclient==2.0.1
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.8
